refactor(Converters): 패키지 이동
- 이전: db/ticket
- 이후: db
