+++ app/src/main/java/woowacourse/movie/db/ticket/TicketDatabase.kt
@@ -5,6 +5,7 @@ import androidx.room.Database
 import androidx.room.Room
 import androidx.room.RoomDatabase
 import androidx.room.TypeConverters
+import woowacourse.movie.db.Converters
 
 @Database(entities = [TicketEntity::class], version = 1, exportSchema = false)
 @TypeConverters(Converters::class)
